+++ Domain/Models/Order/Discount.cs
@@ -1,21 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Domain.Models.Order
 {
     public class Discount
     {
 
-        public Discount(Guid id, double amount)
+        public Discount(Guid id, decimal amount)
         {
             Id = id;
             Amount = amount;
         }
 
         public Guid Id { get; }
-        public double Amount { get; }
+        public decimal Amount { get; }
     }
 }
